backend/requirements.txt
fastapi
uvicorn
cohere
qdrant-client
psycopg2-binary
python-dotenv
google-generativeai
google-genai